Add simple preprocessing logic - from binary integer files to spike times
The first stage of this preprocessing - merging the channels - does so by simply averaging every channel.
The second stage records times when the absoulte value of the potential goes over an arbitrary threshold of 2000 microvolts.
Additionally, it groups these values so as not to fire dozens of times due to noise moving up and down.