Add two weight-setting implementations, and smoothing functionality.
Add a low-pass filter to smooth the incoming data, in hopes that a
derivative-based setup will become more efficient at picking out the
spikes.

Add some additional functionality for setting initial weights for the
network, and add a trial system using networkx to more closely model the
structure of the input.