Refactor again to utilize classes more
This commit changes the architecture of the application somewhat. First
of all, the Neuron class now has subclasses for input, hidden, and
output neurons. This facilitates unique behavior, such as the limitation
on firing an output neuron multiple times.

Also, this commit introduces a new class for encapsulating a network,
and a new function for running a simulation through a given time period.
The network initialization now takes, among other things, a function
which determines how weights are set; this should make it easier to
create networks with new properties.