Pull out network runthrough into its own function
Take the nearly-identical pieces of code found in the initialization
(running through the network with all inputs spiking at time zero) and
the code in the training loop, and put it in the runNetwork function.

This will allow code changes/explanations to be made in one place
instead of two.

Also, rewrite some expressions here and there to use the alpha function
instead of hardcoding, and use the plus-equals operator in a few more
places.